Add writeChangeFileSync and generateChangeFilePath to rush-lib.
