+++ apps/rush-lib/package.json
@@ -22,6 +22,7 @@
     "@types/node": "6.0.62",
     "@types/z-schema": "3.16.31",
     "fs-extra": "~0.26.7",
+    "git-repo-info": "~1.1.4",
     "jju": "~1.3.0",
     "rimraf": "~2.5.4",
     "semver": "~5.3.0",
