+++ apps/rush-lib/typings/git-repo-info/git-repo-info.d.ts
