+++ apps/rush-lib/typings/tsd.d.ts
@@ -1,5 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
+/// <reference path="git-repo-info/git-repo-info.d.ts" />
 /// <reference path="jju/jju.d.ts" />
 /// <reference path="rimraf/rimraf.d.ts" />
