+++ apps/rush/package.json
@@ -35,7 +35,6 @@
     "builtins": "~1.0.3",
     "colors": "~1.1.2",
     "fs-extra": "~0.26.7",
-    "git-repo-info": "~1.1.4",
     "glob": "~7.0.5",
     "glob-escape": "~0.0.1",
     "inquirer": "~1.2.1",
