+++ common/reviews/api/rush-lib.api.ts
@@ -1,3 +1,11 @@
+// @internal
+export function _generateChangeFilePath(rushConfiguration: RushConfiguration,
+  packageName: string): string;
+
+// @internal
+export function _writeChangeFileSync(rushConfiguration: RushConfiguration,
+  changeFile: IChangeFile): void;
+
 // @public
 class ApprovedPackagesConfiguration {
   public constructor(jsonFilename: string);
